feat: added profile page and links to it
