--- src/app/About/About.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function About() {
-  return (
-    <div>
-      
-    </div>
-  )
-}
