+++ src/app/content/Header/Header.jsx
@@ -2,8 +2,8 @@ import React from "react";
 import "./Header.css";
 import Logo from "../components/Logo.jsx";
 import Search from "../components/Search.jsx";
-import About from "../components/AboutLink.jsx";
 import Contact from "../components/ContactLink.jsx";
+import Profile from "../components/ProfileLink.jsx";
 export default function Header() {
   return (
     <header>
@@ -13,8 +13,8 @@ export default function Header() {
           <Search />
         </div>
         <div className="right">
-          <About />
           <Contact />
+          <Profile />
         </div>
       </div>
     </header>
