--- src/app/content/components/AboutLink.css
@@ -1,14 +1,0 @@
-.about {
-  background-color: white;
-  border: 1px solid black;
-  border-radius: 8px;
-  height: 38px;
-  display: flex;
-  align-items: center;
-  font-weight: 700;
-  margin-left: 5px;
-  padding-left: 12px;
-  padding-right: 12px;
-  font-family: system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
-    Oxygen, Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
-}
