--- src/app/content/components/AboutLink.jsx
@@ -1,24 +1,0 @@
-import React from "react";
-import "./ContactLink.css";
-
-export default function About() {
-  return (
-    <a href="/about">
-      <div className="wrapper">
-        <svg
-          width="21"
-          height="20"
-          viewBox="0 0 21 20"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path
-            d="M10.49 0C4.97 0 0.5 4.48 0.5 10C0.5 15.52 4.97 20 10.49 20C16.02 20 20.5 15.52 20.5 10C20.5 4.48 16.02 0 10.49 0ZM17.42 6H14.47C14.15 4.75 13.69 3.55 13.09 2.44C14.93 3.07 16.46 4.35 17.42 6ZM10.5 2.04C11.33 3.24 11.98 4.57 12.41 6H8.59C9.02 4.57 9.67 3.24 10.5 2.04ZM2.76 12C2.6 11.36 2.5 10.69 2.5 10C2.5 9.31 2.6 8.64 2.76 8H6.14C6.06 8.66 6 9.32 6 10C6 10.68 6.06 11.34 6.14 12H2.76ZM3.58 14H6.53C6.85 15.25 7.31 16.45 7.91 17.56C6.07 16.93 4.54 15.66 3.58 14ZM6.53 6H3.58C4.54 4.34 6.07 3.07 7.91 2.44C7.31 3.55 6.85 4.75 6.53 6ZM10.5 17.96C9.67 16.76 9.02 15.43 8.59 14H12.41C11.98 15.43 11.33 16.76 10.5 17.96ZM12.84 12H8.16C8.07 11.34 8 10.68 8 10C8 9.32 8.07 8.65 8.16 8H12.84C12.93 8.65 13 9.32 13 10C13 10.68 12.93 11.34 12.84 12ZM13.09 17.56C13.69 16.45 14.15 15.25 14.47 14H17.42C16.46 15.65 14.93 16.93 13.09 17.56ZM14.86 12C14.94 11.34 15 10.68 15 10C15 9.32 14.94 8.66 14.86 8H18.24C18.4 8.64 18.5 9.31 18.5 10C18.5 10.69 18.4 11.36 18.24 12H14.86Z"
-            fill="#272A37"
-          ></path>
-        </svg>
-        About
-      </div>
-    </a>
-  );
-}
