+++ src/app/content/components/ContactLink.css
@@ -1,29 +1,14 @@
-.wrapper {
+.about {
+  background-color: white;
   border: 1px solid black;
   border-radius: 8px;
-  background-color: white;
   height: 38px;
   display: flex;
-  justify-content: space-between;
   align-items: center;
   font-weight: 700;
   margin-left: 5px;
   padding-left: 12px;
   padding-right: 12px;
-  text-decoration: none;
   font-family: system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
-  Oxygen, Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
-  color: black;
+    Oxygen, Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
 }
-a{
-  text-decoration: none;
-}
-.wrapper svg{
-  padding-top: 2px;
-  padding-right: 5px;
-}
-.wrapper:hover,
-.wrapper:active {
-  background-color: #f5f5f5;
-  text-decoration: none;
-}
